Update patch set 2
Patch Set 2:

Thats what this init is doing. BUT the problem with it is that the dtb, that is responsible for the SDCard problems will still be the one from the row version. You need to unify the dsds and row dtbs in the kernel in order to get everything working with the row dtb. When that is done this init change is ready to be merged.

Patch-set: 2