Update patch set 2
Patch Set 2:

Okay, so sumire is based on ivy but it has diverged, and that'll become a mess and not be maintainable in the future.

What I meant was now that I've made a kitakami-common, once that is merged for ivy, I would rebase sumire on top of ivy, push any common necessary commits to kitakami.

Also, attribution means that this patch is clearly the same as my ivy change, even the commit message.
In that case, use something like git am or git cherry-pick to preserve the authorship, and also the date and Change-Id and what not.

Patch-set: 2