Update patch set 1
Patch Set 1: Code-Review-1

This is wrong, unless LTE works on both SIMs at the same time.

You should switch the first sim to 2g, and then the second to LTE, and see if that works.

There used to be code which did this automatically on switching the second sim to lte (when the first was on lte), not sure if it's still there.

Patch-set: 1
Reviewer: Chirayu Desai <2195@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1